requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.23
django-mysql==4.14.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysql-connector-python==9.0.0
numpy==1.24.4
phonenumbers==9.0.9
pillow==10.4.0
PyJWT==2.9.0
sqlparse==0.5.3
tflite-runtime==2.5.0
typing_extensions==4.13.2
tzdata==2025.2
